Split introduction text to respective assignments
In the beginning, we don't need to know all that sensor stuff.

The first time we need to know about it is in assignment 2, and there
only the events matter, so that part is added to the beginning of
assignment 2.

The information about the asynchronous getters, and that they can fail,
is only required for the assignment 6, so it's move to that.

The text of assignment 6 is also slightly adapted. We clearly show that
Redux Thunk shall to be used here.